requirements.txt
gunicorn
torch==2.2.2
torchvision==0.17.2
flask
numpy==1.26.4
pillow
scikit-learn
werkzeug
opencv-python-headless
